A new ASF appears!
King Raptor reporting for take off.
We now control the skies.

Basic optimization of raptor unit script completed.

The KR differs from the raptor in that it gains speed during combat, making it a stickier fighter. It is also armed with flak missiles as a first-strike/alpha capability. The pew-pew lasers have been swapped off for cannons. A better sound is probably needed though.