Add in the ORCA Missile
When you absolutely need to overkill everything in area..
Further optimization for subprojectilespawner.

This is meant to be a game ender or game changer. Costing a whopping 6.5k, the destruction this can create is rivaled only by nuke. However, like a nuke, it can be targeted by anti during the first and second stage. This allows it to screen for nukes to some degree.

Known issues: Anti-missile must be updated to check for projectile nonexistent.